Allow horizontal scrolling on code snippets
This prevents the mobile viewport from getting stretched out sideways
when there are long lines in code snippets.

It also introduces an accessibility issue - keyboard users may not be
able to focus the code snippets to scroll them horizontally. To fix this
I'd need to add a tabindex attribute to the pre tags. Sadly this
involves customising the markdown processor, which I don't have the
energy for right now. I'm considering a move from Jekyll in any case -
the GDS tech docs tool has solved a bunch of issues like this, which it
would be good to inherit somehow. It uses middleman instead of jekyll.